Remove incorrect calls to XFree()
The pointers returned by these functions don't point to distinct memory
buffers that should be managed by caller. At least with current
xorg/Xrandr implementation, the return values point to internal fields
of the XRRScreenConfiguration struct that was passed to it, so it will
be reclaimed later with a call to XRRFreeScreenConfigInfo().