Upload docs to Hackage on release
Hackage can't build X11 because it doesn't install the required
libraries and headers (Xinerama, Xrandr, …). Therefore we need to upload
the docs ourselves.

I decided to piggyback on the haskell-ci workflow instead of having a
separate workflow as building the docs is not as trivial as the sdist.
Additionally, the sdist and haddock tarball are published as artifacts
so that people can download snapshots and we could possibly use that for
the xmonad website as well.