ci: Manual workflow dispatch for uploading the final release to Hackage
Publishing a candidate on Hackage discards the uploaded docs:
https://github.com/haskell/hackage-server/issues/70#issuecomment-85201,
so as a workaround I'm extending the workflow with the ability to do the
final release with docs upload. To do that, one needs to trigger the
workflow manually and enter the correct version number. Hopefully that's
enough to prevent making a release by mistake.

Note that this complexity may also become unnecessary if the Hackage
admins install a couple extra deps to make X11 build:
https://github.com/haskell/hackage-server/issues/925