ci: Bump haskell-ci
We add "haddock-components: libs" to cabal.haskell-ci to prevent
haskell-ci from passing --haddock-all to cabal v2-haddock, which builds
docs for all components and the …-docs.tar.gz tarball ends up containing
the docs for the last component, which happens to be tests. We need the
tarball to contain the library docs, for upload to Hackage.